+++ CHANGES.md
@@ -7,6 +7,7 @@ Check also [MIGRATION.md](MIGRATION.md) for possible compatibility problems.
 * [#17] Optimized: `StreamEx.ofTree`, `EntryStream.ofTree`.
 * [#104] Added: `StreamEx.ofReversed()`.
 * [#105] Added: `MoreCollectors.minMax()`.
+* [#108] Added: `IntStreamEx.of(IntBuffer)`, `LongStreamEx.of(LongBuffer)`, `DoubleStreamEx.of(DoubleBuffer)`.
 
 ### 0.6.1